refactor: 리뷰 반영 메뉴관리 컨테이너, 페이지 일원화
